requirements.txt
accelerate==1.4.0
diffusers==0.32.2
gradio==5.21.0
lpips==0.1.4
open_clip_torch==2.31.0
openai==1.96.1
polars==1.24.0
tenacity==9.1.2
tensorboard==2.19.0
torch==2.6.0
torchvision==0.21.0
vision-aided-loss==0.1.0
omegaconf==2.3.0
python-dotenv==1.1.1
transformers==4.49.0
peft==0.14.0
einops==0.8.1
pydantic==2.10.6
opencv-python-headless==4.11.0.86
timm==1.0.15
swanlab
pyiqa